옵션에 따른 상품 장바구니 추가

+++ src/main/java/com/spharosacademy/project/SSGBack/cart/dto/Output/CartOutputDto.java
@@ -20,7 +20,7 @@ public class CartOutputDto {
         private String productName;
         private String productBrand;
         private float price;
-        private int qty;
+        private float qty;
         private String titleImgUrl;
 
         OptionCartOutputDto optionCartOutputDto;
